WIP code to teleport a player into the the dungeon when they strike the door with a key
Now, when a player attacks a block Agro-net checks whether it's a Dungeon door block (we'll use NBT tags for that in the future) and if so, it does some checks like whether they have a shard and shulker box. If they do, it removes the shulker box and TPs them to some test location.

In the future this will call out to DungaDunga so that it can call back to the velocity server (via RCON) to move the player around.