Send server-online event every 15 seconds
Dunga-dunga now automatically cleans up dead instances, but this can
happen due to a network error, or if the dungeon is still starting up
and dunga-dunga attempts to send a player to it.

Instead of just sending the server-online event on startup, we now do so
every 15 seconds so that we don't lose an instance forever if the healt
check just fails once.

Resolves #23